requirements.txt
alembic==0.9.9
aniso8601==3.0.0
atomicwrites==1.1.5
attrs==18.1.0
bcrypt==3.1.4
cffi==1.11.5
click==6.7
eventlet==0.23.0
Flask==1.0.2
Flask-JWT==0.3.2
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
greenlet==0.4.13
gunicorn==19.8.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.3
more-itertools==4.2.0
passlib==1.7.1
pluggy==0.6.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.3
pycparser==2.18
PyJWT==1.4.2
pytest==3.6.1
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
python-slugify==1.2.5
pytz==2018.4
six==1.11.0
SQLAlchemy==1.2.8
Unidecode==1.0.22
urllib3==1.23
webargs==3.0.1
Werkzeug==0.14.1
